requirements.txt
numpy
tensorflow
scipy
pyserial
scikit-learn